// js_modules/dagit/packages/core/src/schedules/types/AllSchedulesRepositoriesFragment.ts
export interface AllSchedulesRepositoriesFragment_RepositoryConnection_nodes {
  __typename: "Repository";
  id: string;
  name: string;
  location: AllSchedulesRepositoriesFragment_RepositoryConnection_nodes_location;
  schedules: AllSchedulesRepositoriesFragment_RepositoryConnection_nodes_schedules[];
}

export interface AllSchedulesRepositoriesFragment_RepositoryConnection {
  __typename: "RepositoryConnection";
  nodes: AllSchedulesRepositoriesFragment_RepositoryConnection_nodes[];
}

export interface AllSchedulesRepositoriesFragment_PythonError_cause {
  __typename: "PythonError";
  message: string;
  stack: string[];
}

export interface AllSchedulesRepositoriesFragment_PythonError {
  __typename: "PythonError";
  message: string;
  stack: string[];
  cause: AllSchedulesRepositoriesFragment_PythonError_cause | null;
}

export type AllSchedulesRepositoriesFragment = AllSchedulesRepositoriesFragment_RepositoryConnection | AllSchedulesRepositoriesFragment_PythonError;